restructured Alto to work with the new OpenADK package
