+++ Alto/Browser/Models/ContextManager+LLMManager/OldContextManager.swift
@@ -1,0 +1,18 @@
+import SwiftSoup
+
+/*
+ class NewContextManager {
+ init() {}
+
+ func pullContextFromPage(for webView: WKWebView) {
+ webView.evaluateJavaScript("document.documentElement.outerHTML.toString()") { result, _ in
+ if let html = result as? String {
+ let dom = NewDOMTree(for: html)
+ print(try? dom.rootElement?.element?.text())
+ } else {
+ print("Failed to Extract Content")
+ }
+ }
+ }
+ }
+ */
