+++ Alto/Browser/Models/ContextManager+LLMManager/OldShortcuts.swift
