+++ Alto/Browser/ViewModels/DragAndDrop/FavoriteDropZoneViewModel.swift
@@ -1,4 +1,7 @@
 //
+import AppKit
+import Observation
+import OpenADK
 
 @Observable
 class FavoriteDropZoneViewModel: DropZoneViewModel {
