+++ Alto/Browser/ViewModels/Topbar/AltoTopBarViewModel.swift
@@ -1,12 +1,10 @@
 //
+import Observation
 
 @Observable
 class AltoTopBarViewModel {
     var state: AltoState
     var topbarState: TopbarState = .hidden
-    var currentTab: AltoTab? {
-        state.browserTabsManager.currentSpace.currentTab
-    }
 
     enum TopbarState {
         case hidden
