+++ Alto/Browser/Views/App/BrowserView.swift
@@ -1,0 +1,26 @@
+//
+
+import OpenADK
+import SwiftUI
+
+struct BrowserView: View {
+    var genaricState: any StateProtocol
+
+    // If you can find a better solution please make a pr!
+    var state: AltoState? {
+        if let altoState = genaricState as? AltoState {
+            return altoState
+        }
+        return nil
+    }
+
+    var body: some View {
+        ZStack {
+            WindowBackgroundView()
+            BrowserContentView()
+        }
+        .environment(state)
+        .preferredColorScheme(PreferencesManager.shared.colorScheme)
+        .ignoresSafeArea()
+    }
+}
