--- Alto/Browser/Views/Browser/BrowserView.swift
@@ -1,13 +1,0 @@
-//
-
-import SwiftUI
-
-struct BrowserView: View {
-    var body: some View {
-        ZStack {
-            WindowBackgroundView()
-            BrowserContentView()
-        }
-        .preferredColorScheme(PreferencesManager.shared.colorScheme)
-    }
-}
