+++ Alto/Browser/Views/Components/MacOSButtonsView.swift
