+++ Alto/Browser/Views/Components/VisualEffectView.swift
