+++ Alto/Browser/Views/DragAndDrop/hoverZoneView.swift
@@ -1,5 +1,6 @@
 //
 import AppKit
+import OpenADK
 import SwiftUI
 
 // MARK: - neadleView
