+++ Alto/Browser/Views/Settings/CacheInfoView.swift
@@ -1,3 +1,4 @@
+import OpenADK
 import SwiftUI
 
 struct CacheInfoView: View {
