+++ Alto/Browser/Views/Settings/SettingsView.swift
@@ -1,4 +1,5 @@
 //
+import OpenADK
 import SwiftUI
 
 // What is needed:
