+++ Alto/Browser/Views/TabSection/FavoriteDropZoneView.swift
@@ -1,5 +1,6 @@
-//
+// FavoriteDropZoneView.swift
 
+import OpenADK
 import SwiftUI
 
 // MARK: - FavoriteDropZoneView
@@ -28,19 +29,23 @@ struct FavoriteDropZoneView: View {
 
             if !model.showEmptyDropIndicator {
                 if !model.isEmpty {
-                    hoverZoneView(model: HoverZoneViewModel(
-                        state: model.state,
-                        tabLocation: model.tabLocation,
-                        placement: .start
-                    ))
+                    if let tabLocation = model.tabLocation as? TabLocation {
+                        hoverZoneView(model: HoverZoneViewModel(
+                            state: model.state,
+                            tabLocation: tabLocation,
+                            placement: .start
+                        ))
+                    }
 
                     ForEach(Array(model.displayedTabs.enumerated()), id: \.element.id) { index, tabItem in
                         AltoTabView(model: TabViewModel(state: model.state, draggingViewModel: model, tab: tabItem))
-                        hoverZoneView(model: HoverZoneViewModel(
-                            state: model.state,
-                            tabLocation: model.tabLocation,
-                            index: index
-                        ))
+                        if let tabLocation = model.tabLocation as? TabLocation {
+                            hoverZoneView(model: HoverZoneViewModel(
+                                state: model.state,
+                                tabLocation: tabLocation,
+                                index: index
+                            ))
+                        }
                     }
                     Spacer()
                 }
