--- Alto/Open-ADK/Core/Models/AltoState/AltoState.swift
@@ -1,32 +1,0 @@
-//
-import Observation
-
-/// AltoState handles the state for each window specificaly
-///
-/// Allows each window to display a diferent view of the tabs
-@Observable
-class AltoState {
-    var data: Alto
-    var sidebar = false
-    var Topbar: AltoTopBarViewModel.TopbarState = .active
-    var browserTabsManager = BrowserTabsManager()
-    var draggedTab: TabRepresentation?
-
-    init() {
-        data = Alto.shared
-        browserTabsManager.state = self
-    }
-
-    func setup(webView: WKWebView) {
-        data.cookieManager.setupCookies(for: webView)
-    }
-
-    func toggleTopbar() {
-        switch Topbar {
-        case .hidden:
-            Topbar = .active
-        case .active:
-            Topbar = .hidden
-        }
-    }
-}
