--- Alto/Open-ADK/Core/Models/ContextManager+LLMManager/ContextManager.swift
@@ -1,16 +1,0 @@
-import SwiftSoup
-
-class NewContextManager {
-    init() {}
-
-    func pullContextFromPage(for webView: WKWebView) {
-        webView.evaluateJavaScript("document.documentElement.outerHTML.toString()") { result, _ in
-            if let html = result as? String {
-                let dom = NewDOMTree(for: html)
-                print(try? dom.rootElement?.element?.text())
-            } else {
-                print("Failed to Extract Content")
-            }
-        }
-    }
-}
