--- Alto/Open-ADK/ObjCBridge/Alto-Bridging-Header.m
@@ -1,3 +1,0 @@
-// This code comes from Beam's system: https://github.com/beamlegacy/beam
-
-#import "WKWebsiteDataStore+Private.h"
